fix: rename Euler-related files
