+++ apps/explorer/src/utils/wrappers/eulerEvc/positions/close/EulerClosePosition.pure.tsx
@@ -4,19 +4,18 @@ import { SupportedChainId } from '@cowprotocol/cow-sdk'
 
 import { useOrderContext } from '../../../../../components/orders/OrderWrapperDetails/OrderWrapperDetails.provider'
 import { VaultAsset } from '../../../../../hooks/euler'
+import { OwnerLink } from '../../components/OwnerLink.pure'
+import { TokenLink } from '../../components/TokenLink.pure'
+import { formatAmount, subaccountNumber } from '../../euler.utils'
 import {
   ArrowSep,
   DirectionLabel,
-  OwnerLink,
   SubInfo,
   TokenAmount,
   TokenBlock,
-  TokenLink,
   TokenSymbol,
   TradeCard,
-  formatAmount,
-  subaccountNumber,
-} from '../../EulerWrappers.styles'
+} from '../EulerPosition.styles'
 
 export interface EulerClosePositionParams {
   owner: string
