+++ apps/explorer/src/utils/wrappers/eulerEvc/positions/collateral/EulerCollateralSwap.pure.tsx
@@ -8,19 +8,18 @@ import { BaseIconTooltipOnClick } from 'components/Tooltip'
 import * as styledEl from './EulerCollateralSwap.styled'
 
 import { VaultAsset } from '../../../../../hooks/euler'
+import { OwnerLink } from '../../components/OwnerLink.pure'
+import { TokenLink } from '../../components/TokenLink.pure'
+import { formatAmount, subaccountNumber } from '../../euler.utils'
 import {
   ArrowSep,
   DirectionLabel,
-  OwnerLink,
   SubInfo,
   TokenAmount,
   TokenBlock,
-  TokenLink,
   TokenSymbol,
   TradeCard,
-  formatAmount,
-  subaccountNumber,
-} from '../../EulerWrappers.styles'
+} from '../EulerPosition.styles'
 
 export interface EulerCollateralSwapParams {
   owner: string
